move socialLinks to common folder
